feat: arquitetura simples para o projeto + melhoria na legibilidade e log da aplicação
