+++ package.json
@@ -13,6 +13,7 @@
     "discord.js": "^14.15.3",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
+    "morgan": "^1.10.0",
     "node-cron": "^3.0.3",
     "nodemailer": "^6.9.14",
     "puppeteer": "^22.12.1"
